pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "measure"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
  "GitPython>=3.1",
  "PyGithub>=2.3",
  "radon>=6.0",
  "rich>=13.7",
  "streamlit>=1.36",
  "requests>=2.32",
  "scikit-learn>=1.4",
  "playwright>=1.46"
]

[project.scripts]
measure = "measure.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
